Create history request. Add sum update on TabLayout

+++ app/src/main/java/shiverawe/github/com/receipt/data/Shop.kt
@@ -1,5 +1,5 @@
 package shiverawe.github.com.receipt.data
 
-data class Shop( val date: Int,
+data class Shop( val date: Long,
                  val place: String,
                  val sum: String)
